+++ tests/it/main.rs
@@ -2,10 +2,11 @@ mod basic;
 mod errors;
 mod generic;
 mod matching_eq;
-mod mismatches;
 mod mixed;
 mod mock_order;
 mod prefix;
+#[cfg(feature = "pretty-print")]
+mod pretty_mismatches;
 mod unmock;
 
 fn main() {}