+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- New default feature: `pretty-print`
 ### Fixed
 - Hygiene issue when mocking a function with an argument called `eval` which can be unmocked.
 - Using `eq!` or `ne!` as part of disjuctive input tuple in `matching!` macro.